Add files via upload
This is basically a problem of recursion. This project is on the famous 0/1 knapsack problem in which we have n objects and each object has its specific weight and a value. we need to pick those objects which gave us overall maximum benefit but the total weight must not exceed than the given limit of weight. At the end we return the maximum benefit that we can get.